Float effort

+++ modules/math/ft_pow.c
@@ -6,15 +6,15 @@
 /*   By: maghayev <maghayev@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/24 20:38:44 by maghayev          #+#    #+#             */
-/*   Updated: 2019/09/24 20:40:05 by maghayev         ###   ########.fr       */
+/*   Updated: 2019/11/24 21:14:03 by maghayev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../headers/libft.h"
 
-int		ft_pow(int num1, int power)
+int64_t		ft_pow(int64_t num1, int power)
 {
-	int		res;
+	int64_t		res;
 
 	res = 1;
 	while (power--)
